# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Space-X-Telescope API",
        "version": "0.1"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email Address",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Password Confirmation",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Register a user.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email address",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Register a user.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "auth"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::user",
                "responses": {
                    "200": {
                        "description": "Register a user.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "passport": []
                    }
                ]
            }
        },
        "/api/capsules": {
            "get": {
                "tags": [
                    "capsule"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\CapsuleController::index",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status Filter",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display a listing of capsules.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "passport": []
                    }
                ]
            }
        },
        "/api/capsules/{capsule_serial}": {
            "get": {
                "tags": [
                    "capsule"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\CapsuleController::show",
                "parameters": [
                    {
                        "name": "capsule_serial",
                        "in": "path",
                        "description": "Capsule Serial",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display a capsule.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}